backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
paho-mqtt==2.1.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0
python-multipart==0.0.12
qrcode[pil]==8.0
pydantic[email]==2.9.0
